Close #293 run sofa rpc on top of x-protocol
The purpose is to run sofa stack microservice inside the mesh network,
and to create and run end to end test cases for them in the sofa-mesh
project.

We do not want to modify pilot component to correctly distribute xds
config to sofa stack based microservices, instead we want to reuse
the existing x-protocol support code base. To reuse the existing
pilot support for x-protocol, sofa rpc should be able to run on top
of x-protocol.